ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
tensorflow==2.15.0
Pillow==10.1.0
python-multipart==0.0.6
